extracting navbar into separate component
